move metadata dialog
