+++ frontend/src/components/index.ts
@@ -107,6 +107,9 @@ import("./data-table").then(({ DataTable }) => {
 import("./crawl-status").then(({ CrawlStatus }) => {
   customElements.define("btrix-crawl-status", CrawlStatus);
 });
+import("./crawl-metadata-editor").then(({ CrawlMetadataEditor }) => {
+  customElements.define("btrix-crawl-metadata-editor", CrawlMetadataEditor);
+});
 
 customElements.define("btrix-alert", Alert);
 customElements.define("btrix-input", Input);
